dodane security

+++ src/main/java/com/application/app/cookbook/CookbookController.java
@@ -1,6 +1,7 @@
 package com.application.app.cookbook;
 
 import com.application.app.recipe.Recipe;
+import com.application.app.security.SecurityService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,21 +14,27 @@ public class CookbookController {
     @Autowired
     private CookbookService cookbookService;
 
+    @Autowired
+    private SecurityService securityService;
+
     @GetMapping(value = "")
     public ResponseEntity<Cookbook> getCookbook(@RequestBody CookbookRequest cookbook) {
         try {
-            Cookbook response = cookbookService.getCookbook(cookbook.getCookbookId());
-            return new ResponseEntity<>(response, HttpStatus.OK);
+            if (securityService.isSecuredGetCookbook(cookbook.getCookbookId()) == true) {
+                Cookbook response = cookbookService.getCookbook(cookbook.getCookbookId());
+                return new ResponseEntity<>(response, HttpStatus.OK);
+            } else {
+                throw new Exception();
+            }
         } catch (Exception er) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
     }
 
     @GetMapping(value = "/{recipeId}")
-    public ResponseEntity<Recipe> getRecipeFromCookbook(@PathVariable Long recipeId,
-                                                        @RequestBody CookbookRequest cookbookRequest) {
+    public ResponseEntity<Recipe> getRecipeFromCookbook(@PathVariable Long recipeId) {
         try {
-            Recipe response = cookbookService.getRecipeFromCookbook(cookbookRequest.getCookbookId(), recipeId);
+            Recipe response = cookbookService.getRecipeFromCookbook(recipeId);
             return new ResponseEntity<>(response, HttpStatus.OK);
         } catch (Exception er) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
@@ -37,8 +44,12 @@ public class CookbookController {
     @DeleteMapping(value = "")
     public ResponseEntity<Cookbook> removeRecipe(@RequestBody CookbookRecipeRequest cookbookRecipeRequest ) {
         try {
-            Cookbook response = cookbookService.removeRecipe(cookbookRecipeRequest.getRecipeId(), cookbookRecipeRequest.getCookbookId());
-            return new ResponseEntity<>(response, HttpStatus.OK);
+            if (securityService.isSecuredRemoveRecipeFromCookbook(cookbookRecipeRequest.getCookbookId()) == true) {
+                Cookbook response = cookbookService.removeRecipe(cookbookRecipeRequest.getRecipeId(), cookbookRecipeRequest.getCookbookId());
+                return new ResponseEntity<>(response, HttpStatus.OK);
+            } else {
+                throw new Exception();
+            }
         } catch (Exception er) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
